backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.5.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
beautifulsoup4==4.12.2
reportlab==4.0.4
fpdf2==2.7.6
gunicorn==21.2.0
waitress==3.0.0
google-generativeai==0.3.2